%[ER:97 ]% Make sure test don't end up in end-less loop by doing port.close() after receiving F_DISCONNECTED_SIG. Check if subband selection doesn't contain duplicate subbands.

+++ trunk/MAC/APL/PAC/BeamServer/test/CheckWeightsTest.cc
@@ -81,7 +81,7 @@ public:
 
     void check_weights()
 	{
-	  set<unsigned short> subbands;
+	  set<int> subbands;
 	  Range all = Range::all();
 
 	  //
